+++ s4e-backend/src/main/java/pl/cyfronet/s4e/bean/OverlayOwner.java
@@ -1,0 +1,19 @@
+package pl.cyfronet.s4e.bean;
+
+public enum OverlayOwner {
+    /**
+    * Available for each user
+    */
+    GLOBAL,
+
+    /**
+     * Created by user and for it's purposes only
+     */
+    PERSONAL,
+
+    /**
+     * Created by institution administrator
+     * and available only for members of the institution
+     */
+    INSTITUTIONAL
+}