+++ s4e-backend/src/main/java/pl/cyfronet/s4e/bean/PRGOverlay.java
@@ -1,12 +1,8 @@
 package pl.cyfronet.s4e.bean;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import javax.persistence.*;
-import javax.validation.constraints.NotEmpty;
 
 @Entity
 @Table(name = "prg_overlay")
@@ -19,9 +15,6 @@ public class PRGOverlay {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @NotEmpty
-    private String name;
-
     /**
      * Must be set to match with the featureTypes in the PRG zip. It will also be the layer name
      */
@@ -31,4 +24,9 @@ public class PRGOverlay {
     private SldStyle sldStyle;
 
     private boolean created;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @EqualsAndHashCode.Exclude
+    @ToString.Exclude
+    private WMSOverlay wmsOverlay;
 }