+++ s4e-backend/src/main/java/pl/cyfronet/s4e/bean/UserPreferences.java
@@ -1,0 +1,17 @@
+package pl.cyfronet.s4e.bean;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+import java.util.List;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserPreferences implements Serializable {
+    private List<Long> nonVisibleOverlays;
+}