+++ s4e-backend/src/main/java/pl/cyfronet/s4e/bean/WMSOverlay.java
@@ -1,12 +1,11 @@
 package pl.cyfronet.s4e.bean;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
+import pl.cyfronet.s4e.bean.audit.CreationAndModificationAudited;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
 
 @Entity
 @Table(name = "wms_overlay")
@@ -14,14 +13,28 @@ import javax.validation.constraints.NotEmpty;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class WMSOverlay {
+public class WMSOverlay extends CreationAndModificationAudited {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @NotEmpty
-    private String name;
+    private String label;
 
-    @NotEmpty
+    @NotNull
     private String url;
+
+    @Enumerated(EnumType.STRING)
+    @NotNull
+    private OverlayOwner ownerType;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @EqualsAndHashCode.Exclude
+    @ToString.Exclude
+    private Institution institution;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @ToString.Exclude
+    @EqualsAndHashCode.Include
+    private AppUser appUser;
 }