+++ s4e-backend/src/main/java/pl/cyfronet/s4e/config/SecurityConfig.java
@@ -84,6 +84,16 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .mvcMatchers(POST, prefix("/invitations/{token}/confirm")).authenticated()
                 .mvcMatchers(PUT, prefix("/invitations/{token}/reject")).permitAll()
 
+                .mvcMatchers(GET, prefix("/overlays")).permitAll()
+                .mvcMatchers(POST, prefix("/overlays/personal")).authenticated()
+                .mvcMatchers(POST, prefix("/overlays/global"))
+                    .access("hasRole('ADMIN')")
+                .mvcMatchers(DELETE, prefix("/overlays/personal/{id}")).authenticated()
+                .mvcMatchers(DELETE, prefix("/overlays/global/{id}"))
+                    .access("hasRole('ADMIN')")
+                .mvcMatchers(PUT, prefix("/overlays/{id}/visible")).authenticated()
+                .mvcMatchers(DELETE, prefix("/overlays/{id}/visible")).authenticated()
+
                 .mvcMatchers(GET, prefix("/institutions/{institution}"))
                     .access("hasRole('ADMIN') || @ish.isMember(#institution)")
                 .mvcMatchers(prefix("/institutions/{institution}", "/institutions/{institution}/**"))