+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/response/OverlayResponse.java
@@ -1,0 +1,17 @@
+package pl.cyfronet.s4e.controller.response;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+@Builder
+public class OverlayResponse {
+    private Long id;
+    private String label;
+    private String url;
+    private String ownerType;
+    private boolean visible;
+    private LocalDateTime createdAt;
+}