+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/response/WMSOverlayResponse.java
@@ -1,9 +1,0 @@
-package pl.cyfronet.s4e.controller.response;
-
-public interface WMSOverlayResponse {
-    Long getId();
-
-    String getName();
-
-    String getUrl();
-}