+++ s4e-web/src/app/views/map-view/state/common.model.ts
@@ -2,7 +2,7 @@ import {ID} from '@datorama/akita';
 
 export interface IUILayer {
   cid: ID;
-  caption: string;
+  label: string;
   active: boolean;
   favourite: boolean;
   isLoading: boolean;